core: make opentelemetry dependencies optional
