+++ core-api/Cargo.toml
@@ -13,7 +13,7 @@ categories = ["development-tools"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-otel_impls = ["opentelemetry"]
+otel_impls = ["dep:opentelemetry"]
 
 [dependencies]
 async-trait = "0.1"
